% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{detect_umap_cols}
\alias{detect_umap_cols}
\title{Detect UMAP columns}
\usage{
detect_umap_cols(lfile, umap_regex_str = "UMAP|umap")
}
\arguments{
\item{lfile}{link to a loom file}

\item{umap_regex_str}{string regex to detect UMAP cols using \code{str_subset}}
}
\value{
a vector of strings of length 2 with names of columns
}
\description{
Detect UMAP columns
}
