% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_get.R
\name{get_genes}
\alias{get_genes}
\title{get_genes}
\usage{
get_genes(
  lfile,
  genes,
  select_layer = "matrix",
  select_row = "Gene",
  select_cols = NA
)
}
\arguments{
\item{lfile}{a connected loom object}

\item{genes}{a vector of strings specifying genes}

\item{select_layer}{a string specifying the
layer to extract data from in the loom file}

\item{select_row}{row in the loom file with gene names}

\item{select_cols}{optional; set obs names}
}
\value{
A tibble with gene expression data
}
\description{
Get genes from a connected loom object
}
