% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{get_genes_from_loom}
\alias{get_genes_from_loom}
\title{Get genes from loom file}
\usage{
get_genes_from_loom(lfile, genes, select_layer = "matrix", select_row = "Gene")
}
\arguments{
\item{lfile}{a connected loom file}

\item{genes}{a vector of strings specifying genes}

\item{select_layer}{a string specifying the
layer to extract data from in the loom file (default: "matrix")}

\item{select_row}{row in the loom file with gene names (default: "Gene")}
}
\value{
A tibble with gene expression data
}
\description{
Get genes from loom file
}
