% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_get.R
\name{get_metadata}
\alias{get_metadata}
\title{get_metadata}
\usage{
get_metadata(lfile, obs = NA)
}
\arguments{
\item{lfile}{a connected loom object}

\item{obs}{(optional) character vector with specified observations; by default will take all}
}
\value{
A tibble with metadata
}
\description{
Get metadata from a connceted loom object
}
