% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_get.R
\name{get_umap_df}
\alias{get_umap_df}
\title{get_umap_df}
\usage{
get_umap_df(lfile, umap_cols = NA, obs = NA, genes = NA, ...)
}
\arguments{
\item{lfile}{a connected loom file}

\item{umap_cols}{specify names of UMAP columns to select; otherwise will be
automatically detected by \code{\link[treadle::detect_umap_cols()]{detect_umap_cols}}}

\item{obs}{a vector strings specifying observation variables}

\item{genes}{a vector of strings specifying genes}

\item{...}{parameters for \code{\link[treadle::get_genes()]{get_genes}}}
}
\value{
A tibble with gene expression data
}
\description{
Get umap coordinates from a connected loom object
}
