% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_gene_violin}
\alias{plot_gene_violin}
\title{Plot gene violin}
\usage{
plot_gene_violin(plot_df, x, gene)
}
\arguments{
\item{plot_df}{a data frame with a categorical variable and gene expression}

\item{x}{categorical variable (symbol)}

\item{gene}{gene expression variable (symbol)}
}
\value{
A ggplot object showing violin plot
}
\description{
Plot gene violin
}
