% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_gene_violin_from_loom}
\alias{plot_gene_violin_from_loom}
\title{Plot gene violin directly from loom file}
\usage{
plot_gene_violin_from_loom(
  lfile,
  genes,
  select_layer = "matrix",
  select_row = "Gene",
  select_cols = "biosample_id",
  subset_cols = NULL,
  x_index = 1,
  ncol = length(genes),
  combine = TRUE
)
}
\arguments{
\item{lfile}{a connected loom file}

\item{genes}{a vector of strings specifying genes to plot}

\item{select_layer}{a string specifying the
layer to extract data from in the loom file (default: "matrix")}

\item{select_row}{row in the loom file with gene names (default: "Gene")}

\item{select_cols}{a string specifying cell-level metadata
to use as a categorical variable. Can be a vector of strings
with additional metadata to attach to the ggplot object
for further manipulation (default: "biosample_id", consistent with the
Single Cell Portal metadata requirements)}

\item{subset_cols}{columns to subset by. NULL by default}

\item{x_index}{the index of select columns to use as the x-axis in the
violin plot (default: 1)}

\item{ncol}{number of columns if multiple genes are specified
(default: length(genes))}

\item{combine}{whether to combine the plots or not.}
}
\value{
A ggplot object showing violin plot
}
\description{
Plot gene violin directly from loom file
}
