% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R, R/plotting_plot.R
\name{plot_umap}
\alias{plot_umap}
\title{Plot umap from data frame}
\usage{
plot_umap(
  umap_df,
  var_str = NA,
  umap_cols = NA,
  pt_size = 2,
  rasterize = TRUE,
  pixels = c(1024, 1024),
  reorder_points = TRUE,
  remove_axis = TRUE,
  rename_axis_labels = TRUE
)

plot_umap(
  umap_df,
  var_str = NA,
  umap_cols = NA,
  pt_size = 2,
  rasterize = TRUE,
  pixels = c(1024, 1024),
  reorder_points = TRUE,
  remove_axis = TRUE,
  rename_axis_labels = TRUE
)
}
\arguments{
\item{umap_df}{a tibble containing umap data}

\item{var_str}{(optional) variable encoded in color of plot, by default will grab the third column in \code{umap_df}}

\item{umap_cols}{(optional) vector of strings specifying umap columns to plot, by default will grab the first two columns in \code{umap_df}}

\item{pt_size}{point size passed to \code{geom_scattermore} or \code{geom_point}}

\item{rasterize}{rasterize with \code{geom_scattermore} or not with \code{geom_point}}

\item{pixels}{number of pixels if rasterize is TRUE}

\item{reorder_points}{whether to order points based on expression}

\item{remove_axis}{removes axis}

\item{plot_df}{A data frame or tibble for plotting}

\item{color_str}{String specifying column name to color the UMAP}

\item{umap1_str}{String specifying column name for x-axis (default: "UMAP_1")}

\item{umap2_str}{String specifying column name for x-axis (default: "UMAP_2")}

\item{legend_pt_size}{Point size for dots in legend (default: 3)}

\item{pt_stroke}{Stroke size for dots in plot (default: 0.1)}

\item{label}{If true, plots label of color_str on plot}

\item{label_text_size}{Size of label text (default: 6)}

\item{drop_na}{Whether or not to drop NAs per the color aesthetic;
also drops rows with "NA" strings}

\item{rasterize_points}{Whether or not to rasterize points}

\item{remove_axis_labels}{changes x and y axis labels to UMAP 1 and UMAP 2}
}
\value{
A ggplot object showing UMAP plot

A ggplot2 object
}
\description{
This functions plots a UMAP from a data frame
}
