% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_var_umap_from_loom}
\alias{plot_var_umap_from_loom}
\title{Plot umap from loom file with metadata}
\usage{
plot_var_umap_from_loom(
  lfile,
  var_str,
  select_cols = NULL,
  umap_df = NULL,
  umap_cols = NULL,
  select_layer = "matrix",
  select_row = "Gene",
  ncol = length(select_cols),
  label = FALSE,
  combine = TRUE,
  drop_na = TRUE,
  pt_size = 0.1,
  pt_stroke = 0.5
)
}
\arguments{
\item{lfile}{a connected loom file}

\item{var_str}{a cell level variable to plot}

\item{select_cols}{a string specifying cell-level metadata
to use as a categorical variable. Can be a vector of strings
with additional metadata to attach to the ggplot object
for further manipulation (default: "biosample_id", consistent with the
Single Cell Portal metadata requirements)}

\item{umap_df}{a data frame with the umap columns. Will override umap_col specification}

\item{umap_cols}{a vector of length 2 specifying the names of the umap columns
umap_cols will be automatically detected if not specified}

\item{select_layer}{layer of the loom file to plot expression data from
(default: "matrix")}

\item{select_row}{row in the loom file with gene names (default: "Gene")}

\item{ncol}{number of columns to plot (default: length(genes))}

\item{label}{whether to label the color variable (default: FALSE;
WARNING, currently there's no check on number of categories, so
be careful, this may be prone to crashing if you use it on numerical
variables)}

\item{combine}{whether to combine the plots}

\item{drop_na}{Whether or not to drop NAs per the color aesthetic;
also drops rows with "NA" strings}

\item{pt_size}{(default: 0.1)}

\item{pt_stroke}{(default: 0.5)}
}
\value{
A ggplot object showing UMAP plot
}
\description{
Plot umap from loom file with metadata
}
