% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_violin}
\alias{plot_violin}
\title{Plot violin}
\usage{
plot_violin(plot_df, x, var)
}
\arguments{
\item{plot_df}{a data frame with a categorical variable and gene expression}

\item{x}{categorical variable on x axis as symbol}

\item{var}{variable as symbol}
}
\value{
A ggplot object showing violin plot
}
\description{
Plot violin
}
