% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scp.R
\name{scp_export_clustering_from_loom}
\alias{scp_export_clustering_from_loom}
\title{Export clustering from loom for the Single Cell Portal}
\usage{
scp_export_clustering_from_loom(
  lfile,
  output_dir,
  filename = "scp_clustering",
  umap_cols = NULL,
  str_subset_regex_group = NA,
  str_subset_regex_numeric = NA,
  include_cells = NULL,
  NAME = NA
)
}
\arguments{
\item{lfile}{a connected loom file}

\item{output_dir}{output directory}

\item{filename}{filename (default: "scp_clustering")}

\item{str_subset_regex_group}{string denoting regular expression to select
group variables from cell-level metadata}

\item{str_subset_regex_numeric}{string denoting regular expression to select
numeric variables from cell-level metadata}

\item{include_cells}{vector of cells to export. By default, includes
all cells (default: NULL)}

\item{NAME}{string specifying the cell IDs to go into the NAME column
(default: NA)}
}
\value{
Nothing. Check the output directory for a new file.
}
\description{
Export clustering from loom for the Single Cell Portal
}
