% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scp.R
\name{scp_export_matrix_from_loom}
\alias{scp_export_matrix_from_loom}
\title{Export matrix from loom for the Single Cell Portal}
\usage{
scp_export_matrix_from_loom(
  lfile,
  output_dir,
  layer_name = "matrix",
  gzip = TRUE,
  NAME = "NAME",
  Gene = "Gene"
)
}
\arguments{
\item{lfile}{a connected loom file}

\item{output_dir}{output directory}

\item{layer_name}{the name of the layer to export. The file will be named
accordingly}

\item{gzip}{whether to use gzip to compress the object (default: TRUE)}

\item{NAME}{string specifying the cell IDs to select (default: "NAME")}

\item{Gene}{string specifying the gene IDs to select (default: "Gene")}
}
\value{
Nothing. Check the output directory for a new file.
}
\description{
Export matrix from loom for the Single Cell Portal
}
