% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scp.R
\name{scp_export_metadata_from_loom}
\alias{scp_export_metadata_from_loom}
\title{Export metadata from loom for the Single Cell Portal}
\usage{
scp_export_metadata_from_loom(
  lfile,
  output_dir,
  filename = "scp_metadata",
  str_subset_regex_group = NA,
  str_subset_regex_numeric = NA,
  fill_metadata = FALSE,
  fill_metadata_cellname = NA
)
}
\arguments{
\item{lfile}{a connected loom file}

\item{output_dir}{output directory}

\item{filename}{filename (default: "scp_metadata")}

\item{str_subset_regex_group}{string denoting regular expression to select
group variables from cell-level metadata}

\item{str_subset_regex_numeric}{string denoting regular expression to select
numeric variables from cell-level metadata}

\item{fill_metadata}{boolean specifying whether or not to fill in
arbitrary metadata. USE WITH CAUTION and don't forget about it.}

\item{fill_metadata_cellname}{Cellname to specify when using fill_metadata option

This function will create a 'tmp.txt' file. If you routinely
name files that... maybe watch out? Will be fixed eventually.}
}
\value{
Nothing. Check the output directory for a new file.
}
\description{
Export metadata from loom for the Single Cell Portal
}
