% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{seurat_add_to_loom}
\alias{seurat_add_to_loom}
\title{Export metadata from loom for the Single Cell Portal}
\usage{
seurat_add_to_loom(
  so,
  lfile,
  clustering_assay = "RNA",
  overwrite = FALSE,
  npcs = 30,
  str_subset_regex_cols = "_snn_res"
)
}
\arguments{
\item{so}{Seurat object}

\item{lfile}{connection to loom object}

\item{clustering_assay}{string specifying clustering assay}

\item{overwrite}{whether or not to overwrite the loom object (default: FALSE)}

\item{npcs}{How many PCs to include (default: 30)}

\item{str_subset_regex_cols}{the string for selecting columns by regular expression (default: "_snn_res")}
}
\value{
nothing, the loom file should be updated
}
\description{
Export metadata from loom for the Single Cell Portal
}
