% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{seurat_add_to_loom_init}
\alias{seurat_add_to_loom_init}
\title{Export metadata from loom for the Single Cell Portal}
\usage{
seurat_add_to_loom_init(
  so,
  filename,
  clustering_assay = "RNA",
  species__ontology_label = "Mus musculus",
  species = NA,
  disease__ontology_label = "diabetes mellitus",
  disease = "MONDO_0005015",
  organ__ontology_label = "skin epidermis",
  organ = "UBERON_0001003",
  library_preparation_protocol__ontology_label = "10x 5' v2",
  library_preparation_protocol = "EFO_0009900"
)
}
\arguments{
\item{so}{Seurat object}

\item{clustering_assay}{string specifying clustering assay}

\item{species__ontology_label}{string specifying species__ontology_label required by the Single Cell Portal (default: "Mus musculus")}

\item{species}{string specifying species required by the Single Cell Portal (default: NA)}

\item{disease__ontology_label}{etc. (default: "diabetes mellitus")}

\item{disease}{etc. (default: "MONDO_0005015")}

\item{organ__ontology_label}{etc. (default: "skin epidermis")}

\item{organ}{etc. (default: "UBERON_0001003")}

\item{library_preparation_protocol__ontology_label}{etc. (default: "10x 5' v2")}

\item{library_preparation_protcol}{etc. (default: "EFO_0009900")}
}
\value{
connection to the newly created loom file
}
\description{
Export metadata from loom for the Single Cell Portal
}
