% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{seurat_to_loom}
\alias{seurat_to_loom}
\title{seurat_to_loom}
\usage{
seurat_to_loom(so, filename)
}
\arguments{
\item{so}{Seurat object}

\item{filename}{path to file}
}
\value{
connection to the newly created loom file
}
\description{
Creates a loom function from a Seurat object
}
