% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_umap}
\alias{plot_umap}
\title{Plot umap from data frame}
\usage{
plot_umap(
  plot_df,
  color_str = NULL,
  umap1_str = "UMAP_1",
  umap2_str = "UMAP_2",
  legend_pt_size = 3,
  pt_size = 0.1,
  pt_stroke = 0.5,
  label = FALSE,
  label_text_size = 6,
  reorder_points = TRUE,
  drop_na = TRUE
)
}
\arguments{
\item{plot_df}{A data frame or tibble for plotting}

\item{color_str}{String specifying column name to color the UMAP}

\item{umap1_str}{String specifying column name for x-axis (default: "UMAP_1")}

\item{umap2_str}{String specifying column name for x-axis (default: "UMAP_2")}

\item{legend_pt_size}{Point size for dots in legend (default: 3)}

\item{pt_size}{Point size for dots in plot (default: 0.1)}

\item{pt_stroke}{Stroke size for dots in plot (default: 0.5)}

\item{label}{If true, plots label of color_str on plot}

\item{label_text_size}{Size of label text (default: 6)}

\item{reorder_points}{Whether or not to order the points by increasing
values by the color variable (default: TRUE)}

\item{drop_na}{Whether or not to drop NAs per the color aesthetic;
also drops rows with "NA" strings}
}
\value{
A ggplot object showing UMAP plot
}
\description{
Plot umap from data frame
}
